backend/requirements.txt
Django==4.0.2
django_environ==0.8.1
djangorestframework==3.13.1
environ==1.0
Pillow==9.0.1
django-cors-headers==3.11.0
